Make slate-testing-library its own package
