+++ docs/introduction.mdx
@@ -24,7 +24,11 @@ Coming soon!
 
 ## Testing
 
-Coming soon!
+### [slate-testing-library](/slate-testing-library)
+
+**Note: This library is experimental and may change**
+
+A library to help with writing tests for SlateJS editors.
 
 ## Development
