+++ doczrc.js
@@ -4,7 +4,7 @@ export default {
   title: "ConvertKit Slate Plugins",
   description:
     "A collection of plugins for SlateJS maintained by the ConvertKit team.",
-  menu: ["Development"],
+  menu: ["Introduction", "Development", "Plugins", "Slate Testing Library"],
   plugins: [
     css({
       preprocessor: "postcss"
