+++ packages/slate-testing-library/__tests__/index.js
@@ -1,0 +1,5 @@
+describe("SlateTest", () => {
+  it("should work ", () => {
+    expect(true).toBe(true);
+  });
+});
