+++ packages/slate-testing-library/babel.config.js
@@ -1,0 +1,1 @@
+module.exports = require("../../babel.config.js");
